+++ lib/src/lib/alignment.dart
@@ -1,1 +1,0 @@
-enum ToastAlignment { TOP, BOTTOM, CENTER }