+++ lib/src/lib/stack_toast_manager.dart
@@ -1,58 +1,99 @@
 import 'package:flutter/widgets.dart';
 
+import './simple_toast_view.dart';
 import './stack_toast_view.dart';
 
-class StackToastManager {
-  static StackToastManager? _singleton;
+class FlutterToastManager {
+  static FlutterToastManager? _singleton;
 
-  factory StackToastManager(BuildContext context) {
-    _singleton ??= StackToastManager._internal(context);
+  factory FlutterToastManager(BuildContext context) {
+    _singleton ??= FlutterToastManager._internal(context);
     return _singleton!;
   }
 
   OverlayState? _overlay;
-  StackToastView? _toastView;
-  GlobalKey<StackToastViewState> widgetKey = GlobalKey<StackToastViewState>();
 
-  OverlayEntry? _overlayEntry;
+  OverlayEntry? _simpleOverlayEntry;
+  SimpleToastView? _simpleToastView;
+  final GlobalKey<SimpleToastViewState> _simpleToastKey = GlobalKey<SimpleToastViewState>();
 
-  StackToastManager._internal(BuildContext context) {
+  StackToastView? _stackToastView;
+  OverlayEntry? _stackOverlayEntry;
+  final GlobalKey<StackToastViewState> _stackToastKey = GlobalKey<StackToastViewState>();
+
+  FlutterToastManager._internal(BuildContext context) {
     _overlay ??= Overlay.of(context);
   }
 
-  void init() {
-    if (_overlayEntry == null) {
-      _overlayEntry = OverlayEntry(builder: (BuildContext context) {
-        return _toastView!;
+  void initSimpleView() {
+    if (_simpleOverlayEntry == null) {
+      _simpleOverlayEntry = OverlayEntry(builder: (BuildContext context) {
+        return _simpleToastView!;
       });
-      _overlay?.insert(_overlayEntry!);
+      _overlay?.insert(_simpleOverlayEntry!, above: _stackOverlayEntry);
     }
   }
 
-  void addItem({Widget? widget, Text? text}) {
-    if (widgetKey.currentState == null) {
-      _toastView ??= StackToastView(
+  void addItemToSimpleToast(Duration duration, {Widget? widget, Text? text}) {
+    if (_simpleToastKey.currentState == null) {
+      _simpleToastView ??= SimpleToastView(
         () {
-          _overlayEntry?.remove();
-          _overlayEntry = null;
-          _toastView = null;
+          _simpleOverlayEntry?.remove();
+          _simpleOverlayEntry = null;
+          _simpleToastView = null;
         },
         firstCallAfterBuild: () {
-          widgetKey.currentState?.addWidget(view: widget, text: text);
+          _simpleToastKey.currentState?.addWidget(duration, view: widget, text: text);
         },
-        key: widgetKey,
+        key: _simpleToastKey,
       );
-      init();
+      initSimpleView();
     } else {
-      widgetKey.currentState?.addWidget(view: widget, text: text);
+      _simpleToastKey.currentState?.addWidget(duration, view: widget, text: text);
     }
   }
 
-  void removeLast() {
-    widgetKey.currentState?.removeLast();
+  void removeFirstFromSimple() {
+    _simpleToastKey.currentState?.removeFirst();
   }
 
-  void clear() {
-    widgetKey.currentState?.clear();
+  void clearFromSimple() {
+    _simpleToastKey.currentState?.clear();
+  }
+
+  void initStackView() {
+    if (_stackOverlayEntry == null) {
+      _stackOverlayEntry = OverlayEntry(builder: (BuildContext context) {
+        return _stackToastView!;
+      });
+      _overlay?.insert(_stackOverlayEntry!, below: _simpleOverlayEntry);
+    }
+  }
+
+  void addItemToStackToast({Widget? widget, Text? text}) {
+    if (_stackToastKey.currentState == null) {
+      _stackToastView ??= StackToastView(
+        () {
+          _stackOverlayEntry?.remove();
+          _stackOverlayEntry = null;
+          _stackToastView = null;
+        },
+        firstCallAfterBuild: () {
+          _stackToastKey.currentState?.addWidget(view: widget, text: text);
+        },
+        key: _stackToastKey,
+      );
+      initStackView();
+    } else {
+      _stackToastKey.currentState?.addWidget(view: widget, text: text);
+    }
+  }
+
+  void removeLastFromStack() {
+    _stackToastKey.currentState?.removeLast();
+  }
+
+  void clearFromStack() {
+    _stackToastKey.currentState?.clear();
   }
 }