+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.2
+
+* Add show simple toast
+* Add alignment
+* bug fixed
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Show stack toast view